stockNo now reduces by unitsPerDosage when reminder is selected

+++ project/lib/prescriptions_files/prescriptionCard.dart
@@ -7,7 +7,7 @@ class PrescriptionCard extends StatelessWidget {
   String name;
   double strength;
   String strengthUnits;
-  double unitsPerDosage;
+  int unitsPerDosage;
   String reminderFreq;
   int daysInterval;
   List reminderTimes;
