+++ project/lib/prescriptions_files/prescriptionPopup.dart
@@ -13,7 +13,7 @@ class PrescriptionPopup extends StatelessWidget{
   String name;
   double strength;
   String strengthUnits;
-  double unitsPerDosage;
+  int unitsPerDosage;
   String reminderFreq;
   int daysInterval;
   List reminderTimes;
